+++ aitune/utils/monitoring/__init__.py
@@ -6,13 +6,22 @@ from aitune.utils.monitoring.annotation import annotate
 from aitune.utils.monitoring.context.backend_context import with_backend_context
 from aitune.utils.monitoring.hardware_metrics_annotation import collect_hardware_metrics
 from aitune.utils.monitoring.nvtx_annotation import annotate_with_nvtx
-from aitune.utils.monitoring.setup_hardware_metrics import dump_metrics, get_hardware_metrics
+from aitune.utils.monitoring.setup_hardware_metrics import (
+    disable_hardware_metrics,
+    dump_metrics,
+    enable_hardware_metrics,
+    get_hardware_metrics,
+    snapshot,
+)
 
 __all__ = [
     "annotate",
     "annotate_with_nvtx",
     "collect_hardware_metrics",
+    "disable_hardware_metrics",
     "dump_metrics",
+    "enable_hardware_metrics",
     "get_hardware_metrics",
+    "snapshot",
     "with_backend_context",
 ]