+++ aitune/utils/monitoring/hardware_metrics_annotation.py
@@ -4,28 +4,26 @@
 
 from functools import wraps
 
-from aitune.utils.monitoring.hardware_metrics.collector import NoOpHardwareMetricsCollector
 from aitune.utils.monitoring.setup_hardware_metrics import get_default_collector
 
 
 class collect_hardware_metrics:  # noqa: N801
     """Decorator/context manager for collecting hardware metrics."""
 
-    def __init__(self, name: str | None = None, collector: NoOpHardwareMetricsCollector | None = None):
+    def __init__(self, name: str | None = None):
         """Initializes the hardware metrics collector."""
         self.name = name
-        self.collector = collector or get_default_collector()
 
     def __enter__(self):
         """Enters the context manager."""
         if self.name is None:
             raise ValueError("name must be provided when using collect_hardware_metrics as a context manager")
-        self.collector.start_scope(self.name)
+        get_default_collector().start_scope(self.name)
         return self
 
     def __exit__(self, exc_type, exc_value, traceback):
         """Exits the context manager."""
-        self.collector.end_scope()
+        get_default_collector().end_scope()
         return False
 
     def __call__(self, func):
@@ -34,11 +32,11 @@ class collect_hardware_metrics:  # noqa: N801
 
         @wraps(func)
         def wrapper(*args, **kwargs):
-            self.collector.start_scope(name)
+            get_default_collector().start_scope(name)
             try:
                 result = func(*args, **kwargs)
             finally:
-                self.collector.end_scope()
+                get_default_collector().end_scope()
             return result
 
         return wrapper